Corrected the label of the fahrenheit button to kelvin but also added a fahrenheit button to the fragment_settings.xml
Updated the HomePageFragment.kt so that it displays the correct temperature based on what the user sets within the settings page.

Started on the backend on the SettingsFragment.kt for the temperature conversion logic and storing the users preferred temperature setting.

Added the Kelvin string name into the strings.xml