Created a new branch called Dropdown_Menu to work on adding our app to the drop down menu on a device.
Created a new kotlin class called SkySphereTileService.kt which carries out the function for the Dropdown_Menu branch.

Added the necessary coding to the AndroidManifest.xml file for the tile services.

Also changed the icon of our app.

Although I have started the implementation for the Tile Services feature, it still has errors and is not functioning properly.