Created a TDD test for CalendarManager.kt which is not the same as the attempted TDD test I had made previously and undid.
Made the method of deleteEvent and getEventsByTitleAndTime public so that the CalendarManagerTest.kt could access the methods and made the CalendarManagerTest.kt class open for my 3rd TDD test.

Auto added necessary changes to other files in order for the TDD test to work.

Have written and resolved all 3 of my TDD tests for my CalendarManagerTest.kt.